requirements.txt
fastapi
uvicorn
numpy
scikit-learn
tensorflow
keras
pillow
matplotlib
python-multipart